Split extract code
